feat: remove start.ts and move start to bin
- Makes imports from external modules a little safer.
